requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.1
requests>=2.32.0
assemblyai>=0.35.0
google-generativeai>=0.7.2
pycountry>=22.3.5
Jinja2>=3.1.4
